+++ build-logic/src/main/kotlin/jp/juggler/konaResource/buildlogic/KonaBuildTarget.kt
@@ -1,23 +1,26 @@
 package jp.juggler.konaResource.buildlogic
 
+import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+import org.jetbrains.kotlin.gradle.plugin.getKotlinPluginVersion
 
 /**
  * Kotlin/Nativeのビルドターゲット
  */
 enum class KonaBuildTarget(
     val targetName: String,
-    private val isMacos: Boolean = false,
+    val konanTargetName: String,
 ) {
-    LinuxX64("linuxX64"),
-    LinuxArm64("linuxArm64"),
-    MingwX64("mingwX64"),
-    MacosArm64("macosArm64", isMacos = true),
+    LinuxX64("linuxX64", "linux_x64"),
+    LinuxArm64("linuxArm64", "linux_arm64"),
+    MingwX64("mingwX64", "mingw_x64"),
+    MacosArm64("macosArm64", "macos_arm64"),
 
     // Note:MacOS X64 は Kotlin Native 2.3.20以降でdeprecated
     ;
 
-    internal fun isAvailable(): Boolean = !isMacos || macosBuildAvailable()
+    internal fun isAvailable(project: Project, kotlinVersion: String): Boolean =
+        konanTargetName in project.availableKotlinNativeTargets(kotlinVersion)
 }
 
 private var cacheAvailableTargets :List<KonaBuildTarget>? = null
@@ -27,16 +30,16 @@ private var cacheAvailableTargets :List<KonaBuildTarget>? = null
  * 現在のビルド環境で利用可能なKotlin/Nativeターゲットを返す。
  */
 @Suppress("unused")
-fun availableKonaBuildTarget(): List<KonaBuildTarget> =
+fun Project.availableKonaBuildTarget(): List<KonaBuildTarget> =
     cacheAvailableTargets?:run{
-        KonaBuildTarget.entries.filter { it.isAvailable() }
+        KonaBuildTarget.entries.filter { it.isAvailable(this, getKotlinPluginVersion()) }
     }.also{ cacheAvailableTargets = it}
 
 /**
  * 利用可能なKotlin/NativeターゲットをKotlin Multiplatformへ登録する。
  */
 fun KotlinMultiplatformExtension.konaTargets() {
-    availableKonaBuildTarget().forEach { target ->
+    project.availableKonaBuildTarget().forEach { target ->
         when (target) {
             KonaBuildTarget.LinuxX64 -> linuxX64()
             KonaBuildTarget.LinuxArm64 -> linuxArm64()